+++ packages/varlet-ui/src/list/List.vue
@@ -26,14 +26,15 @@
 <script lang="ts">
 import VarLoading from '../loading'
 import Ripple from '../ripple'
-import { defineComponent, onUnmounted, ref, nextTick, type Ref } from 'vue'
-import { getParentScroller, toPxNum } from '../utils/elements'
+import { defineComponent, onUnmounted, ref, nextTick, type Ref, onDeactivated, watch } from 'vue'
+import { getParentScroller, getRect, toPxNum } from '../utils/elements'
 import { props } from './props'
 import { isNumber } from '@varlet/shared'
 import { dt } from '../utils/shared'
 import { createNamespace, call } from '../utils/components'
 import { pack } from '../locale'
 import { useMounted } from '@varlet/use'
+import { useTabItem } from './provide'
 
 const { n, classes } = createNamespace('list')
 
@@ -45,6 +46,7 @@ export default defineComponent({
   },
   props,
   setup(props) {
+    const { tabItem, bindTabItem } = useTabItem()
     const listEl: Ref<HTMLElement | null> = ref(null)
     const detectorEl: Ref<HTMLElement | null> = ref(null)
     let scroller: HTMLElement | Window
@@ -56,38 +58,48 @@ export default defineComponent({
     }
 
     const isReachBottom = () => {
-      const containerBottom =
-        scroller === window ? window.innerHeight : (scroller as HTMLElement).getBoundingClientRect().bottom
-
-      const { bottom: detectorBottom } = (detectorEl.value as HTMLElement).getBoundingClientRect()
+      const { bottom: containerBottom } = getRect(scroller)
+      const { bottom: detectorBottom } = getRect(detectorEl.value!)
 
       // The fractional part of the detectorBottom when bottoming out overflows
       // https://github.com/varletjs/varlet/issues/310
       return Math.floor(detectorBottom) - toPxNum(props.offset) <= containerBottom
     }
 
+    const removeScrollerListener = () => {
+      scroller.removeEventListener('scroll', check)
+    }
+
     // expose
     const check = async () => {
       await nextTick()
 
-      const { loading, finished, error } = props
-
-      if (!loading && !finished && !error && isReachBottom()) {
-        load()
+      if (props.loading || props.finished || props.error || tabItem?.current.value === false || !isReachBottom()) {
+        return
       }
+
+      load()
     }
 
+    call(bindTabItem, {})
+
+    if (tabItem) {
+      watch(() => tabItem.current.value, check)
+    }
+
+    watch(() => [props.loading, props.error, props.finished], check)
+
     useMounted(() => {
-      scroller = getParentScroller(listEl.value as HTMLElement)
-
-      props.immediateCheck && check()
-
+      scroller = getParentScroller(listEl.value!)
       scroller.addEventListener('scroll', check)
+
+      if (props.immediateCheck) {
+        check()
+      }
     })
 
-    onUnmounted(() => {
-      scroller.removeEventListener('scroll', check)
-    })
+    onDeactivated(removeScrollerListener)
+    onUnmounted(removeScrollerListener)
 
     return {
       pack,