+++ packages/varlet-ui/src/uploader/props.ts
@@ -7,6 +7,7 @@ export interface VarFile {
   name?: string
   url?: string
   cover?: string
+  progress?: number
   fit?: 'fill' | 'contain' | 'cover' | 'none' | 'scale-down'
   state?: 'loading' | 'success' | 'error'
 }
@@ -71,6 +72,7 @@ export const props = {
     type: Boolean,
     default: false,
   },
+  onBeforeFilter: defineListenerProp<(files: VarFile[]) => Promise<VarFile[]> | VarFile[]>(),
   onBeforeRead: defineListenerProp<(file: VarFile) => Promise<any> | any>(),
   onAfterRead: defineListenerProp<(file: VarFile) => any>(),
   onBeforeRemove: defineListenerProp<(file: VarFile) => any>(),